providers: move `is_base_available()` to providers.py
`is_base_available()` contains charmcraft-specific logic, so it is
separated from the craft-providers interface
